@@ -61,11 +61,11 @@
     "@typescript-eslint/eslint-plugin": "^1.9.0",
     "@typescript-eslint/parser": "^1.9.0",
     "audit-filter": "^0.5.0",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
+    "babel-eslint": "^10.0.2",
+    "babel-loader": "^8.0.6",
     "browserslist": "^4.6.6",
     "caniuse-lite": "^1.0.30000989",
-    "eslint": "^5.16.0",
+    "eslint": "^6.2.1",
     "eslint-config-prettier": "^6.0.0",
     "eslint-plugin-fxa": "^1.0.0",
     "eslint-plugin-react": "^7.14.3",
@@ -103,7 +103,7 @@
     "react-dom": "^16.9.0",
     "react-redux": "^7.1.0-alpha.4",
     "react-router-dom": "^5.0.0",
-    "react-scripts": "3.0.1",
+    "react-scripts": "^3.1.1",
     "react-stripe-elements": "^4.0.0",
     "react-test-renderer": "^16.9.0",
     "redux": "^4.0.4",
@@ -131,10 +131,10 @@
     ],
     "coverageThreshold": {
       "global": {
-        "branches": 84,
-        "functions": 71,
-        "lines": 78,
-        "statements": 78
+        "branches": 94,
+        "functions": 84,
+        "lines": 88,
+        "statements": 88
       }
     }
   },
