@@ -28,10 +28,7 @@ module.exports = {
       // All accounts get this product
       PRODUCT_REGISTERED,
       // Other products come from actual subscriptions
-      // TODO: The FxA DB has a column `productName` that we're using for
-      // product_id. We might want to rename that someday.
-      // https://github.com/mozilla/fxa/issues/1187
-      ...subscriptions.map(({ productName }) => productName),
+      ...subscriptions.map(({ productId }) => productId),
     ];
     // Accounts with at least one subscription get this product
     if (subscriptions.length > 0) {
