@@ -101,8 +101,11 @@ const conf = convict({
   productRedirectURLs: {
     default: {
       '123doneProProduct': 'http://127.0.0.1:8080/',
+      fortressProProduct: 'http://127.0.0.1:9292/download',
+      prod_FUUNYnlDso7FeB: 'https://fortress-latest.dev.lcip.org/',
       prod_Ex9Z1q5yVydhyk: 'https://123done-latest.dev.lcip.org/',
-      prod_FUUNYnlDso7FeB: 'https://123done-stage.dev.lcip.org',
+      // todo get new prod_id for 123done stage
+      prod_FfiuDs9u11ESbD: 'https://123done-stage.dev.lcip.org',
     },
     doc: 'Mapping between product IDs and post-subscription redirect URLs',
     env: 'PRODUCT_REDIRECT_URLS',
